Refactor common service logic to AbstractPollingService
